Safari-611.1.10/LayoutTests/js/throw-type-error-is-unique-expected.txt

Tests ES6 %ThrowTypeError% intrinsic is unique

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS %ThrowTypeError% intrinsic is unique
PASS successfullyParsed is true

TEST COMPLETE
